Moved out MonsterEnums.h
