+++ src/GUI/UI/Houses/TownHall.cpp
@@ -11,7 +11,6 @@
 #include "Engine/Engine.h"
 
 #include "GUI/UI/UIHouses.h"
-#include "GUI/GUIMessageQueue.h"
 #include "GUI/GUIWindow.h"
 #include "GUI/GUIButton.h"
 #include "GUI/GUIFont.h"
